requirements.txt
pyyaml
discord.py
spacy